requirements.txt
ollama
fastapi
uvicorn
psycopg2-binary
pydantic
python-dotenv
json-rpc
pydantic_settings
pydantic[email]